fio: Fix (unsigned) integer overflow issues
This patch tries to address issues found when compiling fio with clang's
undefined behaviour sanitizer on 32 bit Linux. While this generated many
false positives it did identify a number of subtle issues when
rates/amounts/time calcultions went above 32 bit values.

The following job was used:
[global]
time_based
runtime=50m
bs=256k
rw=read
rate=512M
ioengine=null
direct=1
iodepth=8
thread
numjobs=10

[null]
size=1G

This patch should also fix https://github.com/axboe/fio/issues/194 and
https://github.com/axboe/fio/issues/266 because 64 bit Windows is an
LLP64 platform so ints and longs still default to 32 bits.

+++ fio.h
@@ -272,7 +272,7 @@ struct thread_data {
 	unsigned long rate_next_io_time[DDIR_RWDIR_CNT];
 	unsigned long rate_bytes[DDIR_RWDIR_CNT];
 	unsigned long rate_blocks[DDIR_RWDIR_CNT];
-	unsigned long rate_io_issue_bytes[DDIR_RWDIR_CNT];
+	unsigned long long rate_io_issue_bytes[DDIR_RWDIR_CNT];
 	struct timeval lastrate[DDIR_RWDIR_CNT];
 	int64_t last_usec;
 	struct frand_state poisson_state;
